Decoupled cMonster and path recalculation logic, created cPathFinder
